# warehouse-service/src/main/java/com/aqstore/service/warehouse/persistence/entity/Product.java
package com.aqstore.service.warehouse.persistence.entity;

import org.springframework.data.annotation.Id;
import org.springframework.data.relational.core.mapping.Table;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;
import lombok.experimental.SuperBuilder;



@Table("product")
@Data
@SuperBuilder
@AllArgsConstructor
@NoArgsConstructor
@EqualsAndHashCode(callSuper = true)
public class Product extends Auditable{

	@Id
	private Long id;
	private String name;
	private String type;
	private String brand;
	private Integer weight;
	private String description;
    private Stock stock;

}